implementing backwards compatibility for auth_key change, and adding success response to client connection

+++ src/clustering/administration/cli/admin_cluster_link.cc
@@ -2328,8 +2328,8 @@ void admin_cluster_link_t::do_admin_set_auth(const admin_command_parser_t::comma
     auth_semilattice_metadata_t auth_metadata = change_request.get();
 
     std::string key = guarantee_param_0(data.params, "key");
-    if (key.length() > auth_semilattice_metadata_t::max_auth_length) {
-        throw admin_cluster_exc_t(strprintf("the provided authorization key is too long, max length: %zu",
+    if (key.length() > static_cast<size_t>(auth_semilattice_metadata_t::max_auth_length)) {
+        throw admin_cluster_exc_t(strprintf("the provided authorization key is too long, max length: %i",
                                             auth_semilattice_metadata_t::max_auth_length));
     }
     auth_metadata.auth_key.get_mutable() = key;
